rt5645: Add HPO_L voltage fix for headphone detect
The ALC5645 has a tendency to apply too much voltage to the HPO_L line, causing
the GPIO wired into the jack detect to float at 2.0V, preventing jack removal
events from occurring properly on some systems (specifically, the jack detect
interrupt is fired, but when the GPIO is read, since it's floating at 2V it's
"misread" as a plug still being present).

This adds a devicetree variable to adjust the headphone amplifiers to reduce the
voltage, thus allowing headphone jack detect events to function properly.

Change-Id: I68ec5fcb18bb49b91848403a9a3c2d0de61287d9