apex: Add in a PCI-based power controller driver
This driver gives us more direct control over the PCI hotplug state of
the Apex device. Effectively this allows us to keep the device loaded
when a user space program has a file descriptor open to the device, and
automatically removes the root PCI bus from the PCI tree when the file
descriptor is cleaned up.

Change-Id: I251afbfffa6c5d28755a2845db35a4bf76c71f72