apex-power: Factor out power control functions
Moves the meat of the power up/down code into functions that only do
that to improve readability and control throughout the code.
Additionally, this adds a delayed init routine that runs one second
after the module has been initialized to automatically power down the
apex if it is currently on and available.

Change-Id: I9be963375db7d436b39ef30fa9d788441f69f58f