Export Android staging UAPI headers
Android staging headers are normally exported using symlinks
in the Android world, but NXP havenn't included them when
pulling in the Android drivers. In Yocto custom headers are
exposed by building against the STAGING_KERNEL_DIR where the
entire patched kernel resides so there's no real need to export
headers in their build system. For Debian the builddeb script
packages things up into linux-headers-version.deb but is
omitting symlinks.

As a workaround to this messy situation restore the symlinks
and hack them into builddeb. This was we can still track
upstream and package the headers.

Change-Id: If28c1640705762442e552f54a881430f17d7ed3b