Update camera mode switch mechanism
During mode change (resolution, framerate), as long as there's v_blank,
h_blank, and system clk update, the exposure time and gain would need
to be recomputed. The original conditional check was too simple and
incorrect. Since the call only happens at v4l VIDIOC_S_PARM ioctl, we
should be okay of not using the direct mode change at all.

In addition, the AE register need to be turned back on after the
exposure update. This was updated in upstream NxP codebase for a later
ov5647 driver, but never fixed in ov5640 version.

Bug: 144574941
Test: Columbia and Enterprise v4l with various resolutions and changing sequences.

Change-Id: Ic631472c872f550d75868e0867b1a9cbe6d3b698