Add power control for Apex
Creates a character device (apex_power) that when open/closed
toggles the PCIe bus as well as the Apex regulator. This can
be used to signficantly reduce idle power when the TPU is
idle.

Note that this is disabled by default in the phanbell device tree.
To enable the apex-power status must be changed to "okay" and for
full power savings apex-regulators should remove the always-on
property.

Change-Id: I906373b5edf481820b692847ebaba57943b85897